+++ src/views/UserEdit/UserEdit.js
@@ -60,6 +60,7 @@ class UserEdit extends React.Component {
     setTenantId: PropTypes.func,
     isCreateKeycloakUserConfirmationOpen: PropTypes.bool,
     initialAssignedRoleIds: PropTypes.object,
+    isLoadingAffiliationRoles: PropTypes.bool.isRequired,
     checkAndHandleKeycloakAuthUser: PropTypes.func,
     confirmCreateKeycloakUser: PropTypes.func
   }
@@ -483,7 +484,8 @@ class UserEdit extends React.Component {
       setTenantId,
       tenantId,
       setAssignedRoleIds,
-      assignedRoleIds
+      assignedRoleIds,
+      isLoadingAffiliationRoles,
     } = this.props;
 
     const profilePictureConfig = get(resources, 'settings.records[0]');
@@ -525,6 +527,7 @@ class UserEdit extends React.Component {
         tenantId={tenantId}
         setAssignedRoleIds={setAssignedRoleIds}
         assignedRoleIds={assignedRoleIds}
+        isLoadingAffiliationRoles={isLoadingAffiliationRoles}
       />
     );
   }